Fix NPE on getActivity() and code refactoring.
